Moved to an Iterator-style pattern when locating services.
For client-side load balancing purposes, it's important to return not a
List but an Iterator type object. This allows us to change the returned
ServiceInstance at runtime (which is needed for a balanced implementation
of round-robin), and means we don't have to look for service in other
datacenters unless we've tried talking to every instance in our local
datacenter.