Fix shutdown procedure
Instead of shutting down the http client first, which often leads to IllegalStateExceptions because the http connection pool was shut down while ConfigUpdater was retrieving its config, shut down the executor first, and only then the http client.
Also switch to the MoreExecutors shutdown procedure for the executor service, as it means not relying on a custom shutdown procedure.