Only fire a KeyListener change if the value for that key changed
This prevents a lsitener from firing if another key has been changed.
As some listeners might perform tasks which can be reset upon changed
(such as creating a new ThreadPool when the number of threads changes)
we do not want to do this when the old value and the new value are
equal to each other.